chore: BLE states
