--- ts/features/itwallet/presentation/proximity/utils/index.ts
@@ -1,10 +1,0 @@
-import { Linking, Platform } from "react-native";
-import AndroidOpenSettings from "react-native-android-open-settings";
-
-export const openBluetoothPreferences = () => {
-  if (Platform.OS === "ios") {
-    Linking.openURL("App-Prefs:Bluetooth").catch(() => null);
-  } else {
-    AndroidOpenSettings.bluetoothSettings();
-  }
-};
